requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.26
botocore==1.36.26
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
configparser==7.1.0
distro==1.9.0
fastapi==0.115.8
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
lxml==5.3.1
openai==1.63.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
shutils==0.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
